requirements.txt
#Main / Default
Django==4.2.3
django-admin-autocomplete-filter==0.7.1
django-admin-list-filter-dropdown==1.0.3
django-admin-shortcuts==3.0.1
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-queryable-properties
django-shell-plus==1.1.7
djangoql==0.17.1
djangorestframework==3.14.0
psycopg2==2.9.7
python-dotenv
ipython==8.15.0
requests==2.28.1

# JWT Authentication
PyJWT==2.10.1
cryptography==43.0.3


# Documentation
drf-spectacular